+++ src/main.zig
@@ -29,8 +29,7 @@ pub fn main() !void {
     };
     defer res.deinit();
     if (res.args.help) {
-        try stdout.print("Zterm version {s}\nCopyright 2021 by Nathan Fisher\n\n",
-            .{version});
+        try stdout.print("Zterm version {s}\nCopyright 2021 by Nathan Fisher\n\n", .{version});
         usage(0);
     }